+++ application/views/includes/head.php
@@ -25,10 +25,10 @@
   <meta name="twitter:image" content="<?php echo isset($images_output[0])?stripslashes($images_output[0]->image):(isset($images[0])?stripslashes($images[0]->image):site_url('media/img/allRise-logo.gif')); ?>">
   <!--end twitter-->
   <title><?php echo ($title)?$title.' | ':''; ?><?php echo SITE_TITLE; ?></title>
-  <link rel="stylesheet" href="//yui.yahooapis.com/pure/0.3.0/pure-min.css">
   <link rel="stylesheet" href="//netdna.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.css">
   <link rel="stylesheet" href="//fonts.googleapis.com/css?family=Roboto:400,400italic,700,700italic">
   <link rel="stylesheet" href="//ajax.googleapis.com/ajax/libs/jqueryui/1.10.3/themes/ui-lightness/jquery-ui.css">
+  <link rel="stylesheet" href="<?php echo site_url('media/css/pure-min.css'); ?>">
   <link rel="stylesheet" href="<?php echo site_url('media/css/prettyPhoto.css'); ?>">
   <link rel="stylesheet" href="<?php echo site_url('media/css/main.css'); ?>">
   <!--<script src="<?php echo site_url('media/js/vendor/modernizr.min.js'); ?>"></script>-->